pyproject.toml
[tool.poetry]
name = "bss"
version = "0.1.0"
description = ""
authors = ["viktor <bnd-baza@mail.ru>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.68.1"
uvicorn = "^0.15.0"
tortoise-orm = "^0.17.7"
aiomysql = "^0.0.21"
aerich = "^0.5.8"
python-multipart = "^0.0.5"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
